requirements.txt
streamlit
pandas
numpy
newsapi-python
transformers
torch
scikit-learn
yfinance
python-dotenv
matplotlib
plotly
pymongo
wordcloud
requests
mysql-connector-python
tweepy
pytrends
textblob